refactor(gpu_execution_prover): split precomputations.rs into by-concern submodules
Same shape as the tracing.rs split: a directory module with mod.rs plus
4 by-concern files (setup_host, common, unrolled, unified), each item
relocated verbatim and re-exported flat via `pub(crate) use X::*` so every
`crate::precomputations::X` consumer path is unchanged.

Pure code relocation: bodies are byte-identical to 43bf103d's
precomputations.rs modulo the minimum pub(crate) visibility bumps needed
for cross-submodule access (config_logs_for_circuit, build_unified_setup_direct,
UnrolledUnifiedSetup + its fields) and import redistribution. The moved
cpu_unified_setup_direct_matches_cpu_geometry test stays co-located with
build_unified_setup_direct in unified.rs and still runs under its new
module path.
