--- gpu/execution_prover/src/precomputations.rs
@@ -1,578 +1,0 @@
-use gpu_circuit_prover::prover::gkr::setup::GpuGKRSetupHost;
-use gpu_circuit_prover::prover::ProverContext;
-use gpu_circuit_prover::witness::circuit_type::CircuitType;
-use gpu_circuit_prover::witness::circuit_type::UnrolledCircuitType::InitsAndTeardowns;
-use gpu_circuit_prover::witness::circuit_type::{
-    DelegationCircuitType, UnrolledCircuitType, UnrolledMemoryCircuitType,
-    UnrolledNonMemoryCircuitType,
-};
-use gpu_circuit_prover::witness::trace_unrolled::ExecutorFamilyDecoderData;
-use gpu_core::allocator::host::ConcurrentStaticHostAllocator;
-use gpu_core::primitives::field::BF;
-use gpu_core::primitives::machine_type::MachineType;
-use gpu_core::primitives::static_host::{alloc_static_pinned_box_uninit, StaticPinnedBox};
-
-use era_cudart::result::CudaResult;
-
-use crate::upstream::{
-    opcodes_for_full_machine_with_unsigned_mul_div_only_with_mem_word_access_specialization,
-    opcodes_for_reduced_machine, process_binary_into_separate_tables_ext,
-    CSExecutorFamilyDecoderData, CpuGKRSetup, GKRCircuitArtifact, SecurityLevel,
-};
-use std::collections::BTreeMap;
-use std::sync::{Arc, OnceLock};
-use worker::Worker;
-
-/// Lazy GPU-side setup host: the CPU setup and the geometry needed to
-/// produce a `GpuGKRSetupHost` are stored here, and the actual
-/// `GpuGKRSetupHost::precompute_from_cpu_setup` call (which does GPU LDE +
-/// commit + D2H of partial trees and the unified cap) runs on first use
-/// inside the GPU worker, against that worker's already-initialized
-/// `ProverContext`. Uses a `OnceLock` so concurrent workers race once and
-/// the loser drops its Arc instead of installing a duplicate.
-///
-/// The `Arc` wrapper around `GpuGKRSetupHost` is what subsequent workers
-/// clone — every prove() that uses this circuit picks up the same cached
-/// pinned-host buffers.
-pub(crate) struct LazyGpuGKRSetupHost {
-    inner: OnceLock<Arc<GpuGKRSetupHost>>,
-    cpu_setup: Arc<CpuGKRSetup<BF>>,
-    log_lde_factor: u32,
-    log_rows_per_leaf: u32,
-    log_tree_cap_size: u32,
-}
-
-impl LazyGpuGKRSetupHost {
-    pub fn new(
-        cpu_setup: Arc<CpuGKRSetup<BF>>,
-        log_lde_factor: u32,
-        log_rows_per_leaf: u32,
-        log_tree_cap_size: u32,
-    ) -> Self {
-        Self {
-            inner: OnceLock::new(),
-            cpu_setup,
-            log_lde_factor,
-            log_rows_per_leaf,
-            log_tree_cap_size,
-        }
-    }
-
-    /// The CPU-side setup this lazy host wraps. Exposed for program-level
-    /// consumers (`gpu_program_prover`) that recompute setup merkle caps on the
-    /// CPU via `GKRSetup::commit` when assembling a `ProgramProof`.
-    pub fn cpu_setup(&self) -> &Arc<CpuGKRSetup<BF>> {
-        &self.cpu_setup
-    }
-
-    /// First call: builds the host on `context` (one-shot stream sync inside
-    /// `precompute_from_cpu_setup`); subsequent calls return the cached Arc.
-    ///
-    /// Returns `None` when the CPU setup has no columns (e.g.
-    /// InitsAndTeardowns has `generic_lookup_tables_width == 0` and no
-    /// decoder table). The proof flow already treats `setup_transfer` as
-    /// optional, so a column-less circuit simply has no setup to transfer
-    /// or commit.
-    pub fn get_or_init(&self, context: &ProverContext) -> CudaResult<Option<Arc<GpuGKRSetupHost>>> {
-        if self.cpu_setup.hypercube_evals.is_empty() {
-            return Ok(None);
-        }
-        self.inner
-            .get_or_try_init(|| {
-                Ok(Arc::new(GpuGKRSetupHost::precompute_from_cpu_setup(
-                    &self.cpu_setup,
-                    self.log_lde_factor,
-                    self.log_rows_per_leaf,
-                    self.log_tree_cap_size,
-                    context,
-                )?))
-            })
-            .map(|arc| Some(Arc::clone(arc)))
-    }
-}
-
-/// Per-circuit precomputations cached across `prove()` invocations.
-/// `setup_host` is a `LazyGpuGKRSetupHost` populated on first GPU-worker
-/// use; the decoder table is host-only and built eagerly in this
-/// constructor.
-#[derive(Clone)]
-pub(crate) struct CircuitPrecomputations {
-    pub compiled_circuit: Arc<GKRCircuitArtifact<BF>>,
-    pub setup_host: Arc<LazyGpuGKRSetupHost>,
-    pub decoder_host: Option<Arc<StaticPinnedBox<ExecutorFamilyDecoderData>>>,
-}
-
-impl CircuitPrecomputations {
-    /// Build the per-circuit precomputations. The GPU-side `GpuGKRSetupHost`
-    /// is *not* materialized here — `setup_host.get_or_init(context)` does
-    /// that on first use inside a GPU worker.
-    pub fn new(
-        circuit_type: CircuitType,
-        compiled_circuit: GKRCircuitArtifact<BF>,
-        cpu_setup: CpuGKRSetup<BF>,
-        decoder_table_data: Option<&[CSExecutorFamilyDecoderData]>,
-        log_lde_factor: u32,
-        log_rows_per_leaf: u32,
-        log_tree_cap_size: u32,
-    ) -> CudaResult<Self> {
-        assert_eq!(
-            compiled_circuit.trace_len,
-            circuit_type.get_domain_size(),
-            "compiled circuit trace_len disagrees with CircuitType geometry for {circuit_type:?}"
-        );
-        let setup_host = Arc::new(LazyGpuGKRSetupHost::new(
-            Arc::new(cpu_setup),
-            log_lde_factor,
-            log_rows_per_leaf,
-            log_tree_cap_size,
-        ));
-        let decoder_host = match decoder_table_data {
-            Some(rows) if !rows.is_empty() => {
-                let mut buf =
-                    alloc_static_pinned_box_uninit::<ExecutorFamilyDecoderData>(rows.len())?;
-                for (slot, src) in buf.iter_mut().zip(rows.iter().copied()) {
-                    *slot = src.into();
-                }
-                Some(Arc::new(buf))
-            }
-            _ => None,
-        };
-        Ok(Self {
-            compiled_circuit: Arc::new(compiled_circuit),
-            setup_host,
-            decoder_host,
-        })
-    }
-}
-
-/// Build the binary-independent precomputations: every delegation circuit
-/// and inits-and-teardowns. CPU-only — no GPU context needed; the
-/// `GpuGKRSetupHost` for each circuit is materialized lazily on first GPU
-/// worker use.
-///
-/// `whir_logs_for_circuit` is invoked per-`CircuitType` and must return
-/// `(log_lde_factor, log_rows_per_leaf, log_tree_cap_size)` matching the
-/// schedule the GPU worker will use at `prove()` time. Different circuits
-/// use different WHIR schedules, so a single global triple is wrong for the
-/// shared map.
-pub(crate) fn get_common_precomputations<F>(
-    whir_logs_for_circuit: F,
-    worker: &Worker,
-) -> CudaResult<BTreeMap<CircuitType, CircuitPrecomputations>>
-where
-    F: Fn(CircuitType) -> (u32, u32, u32),
-{
-    let mut out = BTreeMap::new();
-    for delegation_type in DelegationCircuitType::get_all_delegation_types()
-        .iter()
-        .copied()
-    {
-        let setup = match delegation_type {
-            DelegationCircuitType::BigIntWithControl => {
-                crate::upstream::get_bigint_with_control_circuit_setup(true, worker)
-            }
-            DelegationCircuitType::Blake2WithCompression => {
-                crate::upstream::get_blake2_with_compression_circuit_setup(true, worker)
-            }
-            DelegationCircuitType::Blake2GFunction => {
-                crate::upstream::get_blake2_g_function_circuit_setup(true, worker)
-            }
-            DelegationCircuitType::KeccakSpecial5 => {
-                crate::upstream::get_keccak_special5_circuit_setup(true, worker)
-            }
-        };
-        let circuit_type = CircuitType::Delegation(delegation_type);
-        let (log_lde_factor, log_rows_per_leaf, log_tree_cap_size) =
-            whir_logs_for_circuit(circuit_type);
-        let precomp = CircuitPrecomputations::new(
-            circuit_type,
-            setup.compiled_circuit,
-            setup.setup,
-            None,
-            log_lde_factor,
-            log_rows_per_leaf,
-            log_tree_cap_size,
-        )?;
-        out.insert(circuit_type, precomp);
-    }
-    let it_setup = crate::upstream::inits_and_teardowns_circuit_setup::<
-        ConcurrentStaticHostAllocator,
-    >(true, worker);
-    let circuit_type = CircuitType::Unrolled(InitsAndTeardowns);
-    let (log_lde_factor, log_rows_per_leaf, log_tree_cap_size) =
-        whir_logs_for_circuit(circuit_type);
-    let precomp = CircuitPrecomputations::new(
-        circuit_type,
-        it_setup.compiled_circuit,
-        it_setup.setup,
-        None,
-        log_lde_factor,
-        log_rows_per_leaf,
-        log_tree_cap_size,
-    )?;
-    out.insert(circuit_type, precomp);
-    Ok(out)
-}
-
-pub(crate) fn build_unrolled_circuit_precomputation(
-    machine_type: MachineType,
-    circuit_type: UnrolledCircuitType,
-    binary_image: &[u32],
-    text_section: &[u32],
-    worker: &Worker,
-    security_level: SecurityLevel,
-) -> CircuitPrecomputations {
-    let setup = build_unrolled_setup(
-        machine_type,
-        circuit_type,
-        binary_image,
-        text_section,
-        worker,
-    );
-    let (circuit, cpu_setup, decoder_data) = match setup {
-        UnrolledSetup::Memory(s) => (s.compiled_circuit, s.setup, s.decoder_data),
-        UnrolledSetup::NonMemory(s) => (s.compiled_circuit, s.setup, s.decoder_data),
-        UnrolledSetup::Unified(s) => (s.compiled_circuit, s.setup, s.decoder_data),
-    };
-    let circuit_type = CircuitType::Unrolled(circuit_type);
-    let (log_lde_factor, log_rows_per_leaf, log_tree_cap_size) =
-        config_logs_for_circuit(circuit_type, security_level);
-    CircuitPrecomputations::new(
-        circuit_type,
-        circuit,
-        cpu_setup,
-        Some(decoder_data.as_slice()),
-        log_lde_factor,
-        log_rows_per_leaf,
-        log_tree_cap_size,
-    )
-    .unwrap()
-}
-
-pub(crate) fn get_common_precomputations_for_all(
-    worker: &Worker,
-    security_level: SecurityLevel,
-) -> BTreeMap<CircuitType, CircuitPrecomputations> {
-    get_common_precomputations(
-        move |ct| config_logs_for_circuit(ct, security_level),
-        worker,
-    )
-    .unwrap()
-}
-
-fn config_logs_for_circuit(
-    circuit_type: CircuitType,
-    security_level: SecurityLevel,
-) -> (u32, u32, u32) {
-    let prover_config =
-        gpu_circuit_prover::prover::config::prover_config(circuit_type, security_level).expect(
-            "ExecutionProverConfiguration validated GPU security level before precomputation",
-        );
-    (
-        prover_config.lde_factor.trailing_zeros(),
-        prover_config.base_oracles_values_per_leaf.trailing_zeros(),
-        prover_config.cap_size.trailing_zeros(),
-    )
-}
-
-struct UnrolledMemorySetup {
-    compiled_circuit: GKRCircuitArtifact<BF>,
-    setup: CpuGKRSetup<BF>,
-    decoder_data: Vec<CSExecutorFamilyDecoderData>,
-}
-
-struct UnrolledNonMemorySetup {
-    compiled_circuit: GKRCircuitArtifact<BF>,
-    setup: CpuGKRSetup<BF>,
-    decoder_data: Vec<CSExecutorFamilyDecoderData>,
-}
-
-struct UnrolledUnifiedSetup {
-    compiled_circuit: GKRCircuitArtifact<BF>,
-    setup: CpuGKRSetup<BF>,
-    decoder_data: Vec<CSExecutorFamilyDecoderData>,
-}
-
-enum UnrolledSetup {
-    Memory(UnrolledMemorySetup),
-    NonMemory(UnrolledNonMemorySetup),
-    Unified(UnrolledUnifiedSetup),
-}
-
-/// Wraps a per-family memory-circuit setup's `(compiled_circuit, setup)`
-/// pair together with its decoder data, mirroring the shape every
-/// `UnrolledCircuitType::Memory` arm in `build_unrolled_setup` produces.
-fn wrap_memory_setup(
-    compiled_circuit: GKRCircuitArtifact<BF>,
-    setup: CpuGKRSetup<BF>,
-    decoder_data: Vec<CSExecutorFamilyDecoderData>,
-) -> UnrolledSetup {
-    UnrolledSetup::Memory(UnrolledMemorySetup {
-        compiled_circuit,
-        setup,
-        decoder_data,
-    })
-}
-
-/// Non-memory counterpart of [`wrap_memory_setup`]; mirrors every
-/// `UnrolledCircuitType::NonMemory` arm in `build_unrolled_setup`.
-fn wrap_non_memory_setup(
-    compiled_circuit: GKRCircuitArtifact<BF>,
-    setup: CpuGKRSetup<BF>,
-    decoder_data: Vec<CSExecutorFamilyDecoderData>,
-) -> UnrolledSetup {
-    UnrolledSetup::NonMemory(UnrolledNonMemorySetup {
-        compiled_circuit,
-        setup,
-        decoder_data,
-    })
-}
-
-fn build_unrolled_setup(
-    machine_type: MachineType,
-    circuit_type: UnrolledCircuitType,
-    binary_image: &[u32],
-    text_section: &[u32],
-    worker: &Worker,
-) -> UnrolledSetup {
-    // The unified circuit is built directly from the unified cs/prover sources
-    // (mirroring `UnifiedRiscvCircuitOracle::new` + `build_unified_artifact`),
-    // not via the per-family decoder preprocessing below. Short-circuit before
-    // any per-family work runs.
-    if let UnrolledCircuitType::Unified = circuit_type {
-        return UnrolledSetup::Unified(build_unified_setup_direct(
-            binary_image,
-            text_section,
-            worker,
-        ));
-    }
-    use riscv_transpiler::ir::{FullUnsignedMachineDecoderConfig, ReducedMachineDecoderConfig};
-    use std::alloc::Global;
-
-    let supported_csrs: Vec<u16> =
-        DelegationCircuitType::get_delegation_types_for_machine_type(machine_type)
-            .iter()
-            .map(DelegationCircuitType::get_delegation_type_id)
-            .collect();
-    let preprocessing = match machine_type {
-        MachineType::Full | MachineType::FullUnsigned => {
-            process_binary_into_separate_tables_ext::<
-                BF,
-                FullUnsignedMachineDecoderConfig,
-                true,
-                Global,
-            >(
-                text_section,
-                &opcodes_for_full_machine_with_unsigned_mul_div_only_with_mem_word_access_specialization(),
-                common_constants::ROM_WORD_SIZE,
-                &supported_csrs,
-            )
-        }
-        MachineType::Reduced => process_binary_into_separate_tables_ext::<
-            BF,
-            ReducedMachineDecoderConfig,
-            true,
-            Global,
-        >(
-            text_section,
-            &opcodes_for_reduced_machine(),
-            common_constants::ROM_WORD_SIZE,
-            &supported_csrs,
-        ),
-    };
-    let family_idx = match circuit_type {
-        UnrolledCircuitType::Memory(c) => c.get_family_idx(),
-        UnrolledCircuitType::NonMemory(c) => c.get_family_idx(),
-        // `Unified` is handled by the early return above; this per-family path is
-        // for memory/non-memory circuits only.
-        UnrolledCircuitType::InitsAndTeardowns | UnrolledCircuitType::Unified => {
-            panic!("build_unrolled_setup is for memory/non-memory circuits only")
-        }
-    };
-    let table_data = preprocessing
-        .get(&family_idx)
-        .expect("missing decoder data for circuit family")
-        .clone();
-    let decoder_data_from_table: Vec<CSExecutorFamilyDecoderData> = table_data
-        .iter()
-        .map(|el| el.as_ref().copied().unwrap_or_default())
-        .collect();
-    match circuit_type {
-        UnrolledCircuitType::Memory(UnrolledMemoryCircuitType::LoadStoreWordOnly) => {
-            let s = crate::upstream::load_store_word_only_circuit_setup::<Global>(
-                &table_data,
-                binary_image,
-                true,
-                worker,
-            );
-            wrap_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::Memory(UnrolledMemoryCircuitType::LoadStoreSubwordOnly) => {
-            let s = crate::upstream::load_store_subword_only_circuit_setup::<Global>(
-                &table_data,
-                binary_image,
-                true,
-                worker,
-            );
-            wrap_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::NonMemory(UnrolledNonMemoryCircuitType::AddSubLuiAuipcMop) => {
-            let s = crate::upstream::add_sub_lui_auipc_mop_circuit_setup::<Global>(
-                &table_data,
-                true,
-                worker,
-            );
-            wrap_non_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::NonMemory(UnrolledNonMemoryCircuitType::JumpBranchSlt) => {
-            let s =
-                crate::upstream::jump_branch_slt_circuit_setup::<Global>(&table_data, true, worker);
-            wrap_non_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::NonMemory(UnrolledNonMemoryCircuitType::ShiftBinaryCsr) => {
-            let s =
-                crate::upstream::shift_binary_circuit_setup::<Global>(&table_data, true, worker);
-            wrap_non_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::NonMemory(UnrolledNonMemoryCircuitType::MulDivUnsigned) => {
-            let s = crate::upstream::mul_div_unsigned_circuit_setup::<Global>(
-                &table_data,
-                true,
-                worker,
-            );
-            wrap_non_memory_setup(s.compiled_circuit, s.setup, decoder_data_from_table)
-        }
-        UnrolledCircuitType::InitsAndTeardowns | UnrolledCircuitType::Unified => unreachable!(),
-    }
-}
-
-/// Build the unified-circuit setup directly from the unified cs/prover sources,
-/// bit-for-bit mirroring the CPU unified decoder-preprocessing path
-/// and `cs::gkr_circuits::unified_reduced_machine::circuit::build_unified_artifact`
-/// compile path (the source of truth that produced
-/// `unified_reduced_machine_layout_gkr.json`). Unlike the per-family path in
-/// `build_unrolled_setup`, this uses the unified 5-CSR
-/// `UnifiedReducedMachineDecoder` over `ReducedMachineDecoderConfig`
-/// (family 128), NOT the per-family decoder set.
-fn build_unified_setup_direct(
-    binary_image: &[u32],
-    text_section: &[u32],
-    _worker: &Worker,
-) -> UnrolledUnifiedSetup {
-    use crate::upstream::{
-        build_unified_table_driver, create_mem_word_only_special_tables,
-        unified_reduced_machine_circuit_with_preprocessed_bytecode_for_gkr,
-        unified_reduced_machine_table_addition_fn, BasicAssembly, Circuit, GKRCompiler,
-        OpcodeFamilyDecoder, UnifiedReducedMachineDecoder, ROM_SECOND_WORD_BITS, ROM_WORD_SIZE,
-    };
-
-    // Compile the unified artifact identically to
-    // `build_unified_artifact::<BF>(use_caches = true)`
-    // (cs::gkr_circuits::unified_reduced_machine::circuit): register the unified
-    // tables + the mem_word_only AlignedRomRead table (sized with empty bytecode
-    // so `offset_for_decoder_table` accounts for it), build the circuit body, then
-    // compile with inline inits/teardowns at trace_len_log2 = 24.
-    let mut cs = BasicAssembly::<BF>::new();
-    unified_reduced_machine_table_addition_fn(&mut cs);
-    for (table_type, table) in create_mem_word_only_special_tables::<BF, ROM_SECOND_WORD_BITS>(&[])
-    {
-        cs.add_table_with_content(table_type, table);
-    }
-    unified_reduced_machine_circuit_with_preprocessed_bytecode_for_gkr(&mut cs);
-    let (cs_output, _) = cs.finalize();
-    let compiler = GKRCompiler::<BF>::default();
-    let compiled_circuit = compiler.compile_family_circuit_with_inline_inits_and_teardowns(
-        cs_output,
-        ROM_WORD_SIZE,
-        /* num_inits_and_teardowns */ 1,
-        /* trace_len_log2 */ 24,
-        /* caching_is_allowed */ true,
-    );
-
-    // Decoder table via the unified reduced-machine preprocessing (mirrors
-    // setups::unrolled_circuits::unifier_reduced_machine_circuit's
-    // `unified_reduced_machine_circuit_setup`: `UnifiedReducedMachineDecoder`
-    // over `ReducedMachineDecoderConfig` with the 5 supported CSRs). pr-332
-    // removed the `UnifiedRiscvCircuitOracle::new` derivation this used to
-    // mirror; the oracle is now a plain borrow of an externally built table.
-    let decoder_table = {
-        use common_constants::circuit_families::REDUCED_MACHINE_CIRCUIT_FAMILY_IDX;
-        use common_constants::{
-            BIGINT_OPS_WITH_CONTROL_CSR_REGISTER, BLAKE2S_DELEGATION_CSR_REGISTER,
-            BLAKE2S_G_FUNCTION_DELEGATION_CSR_REGISTER, KECCAK_SPECIAL5_CSR_REGISTER,
-            NON_DETERMINISM_CSR,
-        };
-        use riscv_transpiler::ir::ReducedMachineDecoderConfig;
-        use std::alloc::Global;
-        let decoders: Vec<Box<dyn OpcodeFamilyDecoder>> =
-            vec![Box::new(UnifiedReducedMachineDecoder)];
-        const SUPPORTED_CSRS: &[u16] = &[
-            NON_DETERMINISM_CSR as u16,
-            BLAKE2S_DELEGATION_CSR_REGISTER as u16,
-            BIGINT_OPS_WITH_CONTROL_CSR_REGISTER as u16,
-            KECCAK_SPECIAL5_CSR_REGISTER as u16,
-            BLAKE2S_G_FUNCTION_DELEGATION_CSR_REGISTER as u16,
-        ];
-        let mut preprocessing_data =
-            crate::upstream::process_binary_into_separate_tables_ext::<
-                BF,
-                ReducedMachineDecoderConfig,
-                true,
-                Global,
-            >(text_section, &decoders, ROM_WORD_SIZE, SUPPORTED_CSRS);
-        preprocessing_data
-            .remove(&REDUCED_MACHINE_CIRCUIT_FAMILY_IDX)
-            .expect("UnifiedReducedMachineDecoder must produce a family-128 entry")
-    };
-
-    // Real binary-derived AlignedRomRead content lives in the prove-time table
-    // driver (matches the second `create_mem_word_only_special_tables` call inside
-    // `build_unified_table_driver`).
-    let table_driver = build_unified_table_driver::<BF>(binary_image);
-
-    let trace_len = CircuitType::Unrolled(UnrolledCircuitType::Unified).get_domain_size();
-    let setup = CpuGKRSetup::construct(&table_driver, &decoder_table, trace_len, &compiled_circuit);
-
-    let decoder_data: Vec<CSExecutorFamilyDecoderData> = decoder_table
-        .iter()
-        .map(|el| el.as_ref().copied().unwrap_or_default())
-        .collect();
-
-    UnrolledUnifiedSetup {
-        compiled_circuit,
-        setup,
-        decoder_data,
-    }
-}
-
-#[cfg(test)]
-mod unified_setup_tests {
-    use super::*;
-    use crate::upstream::OutputType;
-
-    /// CPU-free geometry oracle: the unified artifact the GPU build-direct path
-    /// compiles must match the shape the cs-side test asserts
-    /// (`compile_unified_reduced_machine_with_inline_inits_and_teardowns` in
-    /// `cs::gkr_circuits::unified_reduced_machine::circuit`).
-    #[test]
-    fn cpu_unified_setup_direct_matches_cpu_geometry() {
-        let worker = Worker::new();
-        // 2^(16+ROM_SECOND_WORD_BITS) bytes of zero bytecode for compile-shape parity.
-        let n = (1usize << (16 + crate::upstream::ROM_SECOND_WORD_BITS)) / 4;
-        let binary = vec![0u32; n];
-        let s = build_unified_setup_direct(&binary, &binary, &worker);
-        assert_eq!(
-            s.compiled_circuit.trace_len,
-            CircuitType::Unrolled(UnrolledCircuitType::Unified).get_domain_size()
-        );
-        assert!(s
-            .compiled_circuit
-            .global_output_map
-            .contains_key(&OutputType::PermutationProduct));
-        assert!(s
-            .compiled_circuit
-            .global_output_map
-            .contains_key(&OutputType::InitsAndTeardownsProduct));
-        assert!(!s.compiled_circuit.memory_layout.teardown_sets.is_empty());
-    }
-}
