+++ gpu/execution_prover/src/precomputations/mod.rs
@@ -1,0 +1,9 @@
+mod common;
+mod setup_host;
+mod unified;
+mod unrolled;
+
+pub(crate) use common::*;
+pub(crate) use setup_host::*;
+pub(crate) use unified::*;
+pub(crate) use unrolled::*;
